requirements.txt
flask
gunicorn
Flask-Failsafe
mysql-connector-python
flask_cors
Pillow
opencv-python-headless
google-cloud-vision==2.7.0
firebase-admin
flask-firebase-admin

